requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
python-multipart
supabase
alembic
sqlalchemy
